rag-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
openai==1.58.1
langchain==0.3.14
langchain-openai==0.3.0
langchain-community==0.3.14
pinecone-client==5.0.1
structlog==24.4.0
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-instrumentation-fastapi==0.50b0
prometheus-fastapi-instrumentator==7.0.2
python-dotenv==1.0.1
